refactor: move WordformSearch to search.py
